// web/src/types/store/setting.d.ts
/**
 * 设置状态类型声明
 */
export type SettingState = {
  title: string
  theme: string
  blogTheme: string
  editorPreviewTheme: string
  editorCodeTheme: string
  tagsView: boolean
  fixedHeader: boolean
  showSettings: boolean
  sidebarLogo: boolean
  topNav: boolean
  dynamicTitle: boolean
}